adding some delete hook to avoid to limit a bit what we put in /tmp TOMEE-6: Arquillian adapter
git-svn-id: https://svn.apache.org/repos/asf/openejb/trunk/openejb@1210265 13f79535-47bb-0310-9956-ffa450edef68

+++ arquillian-tomee/arquillian-tomee-common/src/main/java/org/apache/openejb/arquillian/common/FileUtils.java
@@ -38,7 +38,7 @@ public class FileUtils {
         // no-op
     }
 
-    // Shutdown hook for recurssive delete on tmp directories
+    // Shutdown hook for recursive delete on tmp directories
     static final List<String> delete = new ArrayList<String>();
 
     static {
@@ -61,12 +61,14 @@ public class FileUtils {
     }
 
     public static void delete(File file) {
-        if (file.isDirectory()) {
-            for (File f : file.listFiles()) {
-                delete(f);
+        if (file.exists()) {
+            if (file.isDirectory()) {
+                for (File f : file.listFiles()) {
+                    delete(f);
+                }
             }
-        }
 
-        file.delete();
+            file.delete();
+        }
     }
 }
